test - Creates global conftests for teos
The global conftest shares functionality between unit and e2e to avoid replication
